FIX: Change text and add card - Add card with main info weather - Change text english to spanish - Adjust styles and hover styles
